+++ package.json
@@ -84,7 +84,7 @@
 		"release-it": "^15.6.0",
 		"sentences-per-line": "^0.2.1",
 		"should-semantic-release": "^0.1.0",
-		"typescript": "^5.0.0",
+		"typescript": "^5.6.3",
 		"vitest": "^0.31.1",
 		"yaml-eslint-parser": "^1.2.0"
 	},
