+++ src/tests/features/preferPromiseFnWhenKnown.test.ts
@@ -54,29 +54,34 @@ export const Game = {
 	const { vfsMap } = await getDTSFilesForRun({ sdl, gamesService, prismaSchema })
 	const dts = vfsMap.get("/types/games.d.ts")!
 	expect(dts.trim()).toMatchInlineSnapshot(`
-		"interface GameSyncResolver {
-		  (args?: RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>), obj?: RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>)): RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>);
+		"/*SDL: gameSync: Game*/
+		export interface GameSyncResolver {
+		  (args?: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>);
 		}
-		interface GameAsyncResolver {
-		  (args?: Promise<RTGame| null>, obj?: Promise<RTGame| null>): Promise<RTGame| null>;
+		/*SDL: gameAsync: Game*/
+		export interface GameAsyncResolver {
+		  (args?: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): Promise<RTGame| null>;
 		}
-		interface GameAsync1ArgResolver {
-		  (args: RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>), obj?: RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>)): RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>);
+		/*SDL: gameAsync1Arg: Game*/
+		export interface GameAsync1ArgResolver {
+		  (args: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>);
 		}
-		interface GameAsync2ArgResolver {
-		  (args: RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>), obj: RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>)): RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>);
+		/*SDL: gameAsync2Arg: Game*/
+		export interface GameAsync2ArgResolver {
+		  (args: object, obj: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame| null | Promise<RTGame| null> | (() => Promise<RTGame| null>);
 		}
-		interface GameObjResolver {
-		  (args?: RTGame| null, obj?: RTGame| null): RTGame| null;
+		/*SDL: gameObj: Game*/
+		export interface GameObjResolver {
+		  (args?: object, obj?: { root: Query, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }): RTGame| null;
 		}
-		interface GameTypeResolvers<Extended> {
-		  /*SDL: summary: String!*/
+		export interface GameTypeResolvers<Extended> {
+		  /* SDL: summary: String!*/
 		  summary: string;
-		  /*SDL: summarySync: String!*/
+		  /* SDL: summarySync: String!*/
 		  summarySync: (args?: undefined, obj?: { root: GameAsParent<Extended>, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }) => string;
-		  /*SDL: summarySyncBlock: String!*/
+		  /* SDL: summarySyncBlock: String!*/
 		  summarySyncBlock: (args?: undefined, obj?: { root: GameAsParent<Extended>, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }) => string | Promise<string> | (() => Promise<string>);
-		  /*SDL: summaryAsync: String!*/
+		  /* SDL: summaryAsync: String!*/
 		  summaryAsync: (args?: undefined, obj?: { root: GameAsParent<Extended>, context: RedwoodGraphQLContext, info: GraphQLResolveInfo }) => Promise<string>;
 		}
 		type GameAsParent<Extended> = PGame & {summary: () => string,
